requirements.txt
# Core ML/NLP libraries
torch>=2.0.0
transformers>=4.30.0
datasets>=2.14.0
accelerate>=0.20.0
sentence-transformers>=2.2.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.65.0
jsonlines>=3.1.0

# Evaluation and metrics
scikit-learn>=1.3.0
nltk>=3.8.0
rouge-score>=0.1.2

# RAG and retrieval
faiss-cpu>=1.7.4
rank-bm25>=0.2.2

# Utilities
wandb>=0.15.0
matplotlib>=3.7.0
seaborn>=0.12.0
click>=8.1.0

# Development
pytest>=7.4.0
black>=23.7.0
flake8>=6.0.0

# Optional GPU support (comment out if using CPU)
# torch-audio>=2.0.0
# faiss-gpu>=1.7.4